requirements.txt
async-timeout
bidict
click
Deprecated
dnspython
eventlet
Flask
Flask-SocketIO
Flask-WTF
greenlet
itsdangerous
Jinja2
MarkupSafe
packaging
pyparsing
python-engineio
python-socketio
six
typing_extensions
Werkzeug
wrapt
WTForms
redis[hiredis]